requirements.txt
streamlit
streamlit-option-menu
requests
python-dotenv
openai